+++ src/main/java/com/example/gamemate/domain/notification/service/NotificationService.java
@@ -5,13 +5,10 @@ import com.example.gamemate.domain.notification.entity.Notification;
 import com.example.gamemate.domain.notification.enums.NotificationType;
 import com.example.gamemate.domain.notification.repository.NotificationRepository;
 import com.example.gamemate.domain.user.entity.User;
-import com.example.gamemate.domain.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
@@ -25,7 +22,7 @@ import java.util.stream.Collectors;
 public class NotificationService {
 
     private final NotificationRepository notificationRepository;
-    private final JavaMailSender javaMailSender;
+    private final AsyncNotificationService asyncNotificationService;
 
     // 알림 생성
     @Transactional
@@ -47,8 +44,9 @@ public class NotificationService {
     }
 
     // 알림 발송 (이메일)
-    @Scheduled(cron = "0 0/10 * * * *")
-    public void sendNotificationMail() {
+    @Scheduled(cron = "0 0/3 * * * *")
+    public void scheduleNotificationEmail() {
+        log.info("스케쥴링 활성화");
 
         List<Notification> unnotifiedNotificationList = notificationRepository.findAllBySentStatus(false);
 
@@ -66,27 +64,7 @@ public class NotificationService {
             User user = entry.getKey();
             List<Notification> notifications = entry.getValue();
 
-            try {
-                SimpleMailMessage simpleMailMessage = new SimpleMailMessage();
-                simpleMailMessage.setTo(user.getEmail()); // 보낼 사람
-                simpleMailMessage.setSubject("[GameMate] 새로운 알림이 있습니다."); // 제목
-                simpleMailMessage.setFrom("newbiekk1126@gmail.com"); // 보내는 사람
-                simpleMailMessage.setText("새로운 알림이 " + notifications.size() + "개 있습니다."); // 내용
-
-                javaMailSender.send(simpleMailMessage);
-                log.info("{}님에게 {}개의 알림 메일을 전송했습니다.", user.getEmail(), notifications.size());
-
-                updateNotificationStatus(notifications);
-            } catch (Exception e) {
-                log.error("알림 메일 전송 실패: {}", user.getEmail(), e);
-            }
+            asyncNotificationService.sendNotificationMail(user, notifications);
         }
-    }
-
-    // 알림 전송 후 notified(false -> true) 상태 변경
-    @Transactional
-    public void updateNotificationStatus(List<Notification> notifications) {
-        notifications.forEach(notification -> notification.updateSentStatus(true));
-        notificationRepository.saveAll(notifications);
     }
 }
